Addition of classes necessary to implement rendering api as discussed in design session.
Most importantly AuctionHouse, an application scope class that will initializes the application and AuctionitemBean which is a combination of AuctionItem and Bid that is used to represent an item in the UI.  The AuctionitemBean will make render calls when necessary.

Other classes have been added to mirror the Seam implementation in the Booking demo.