Correct implementation of Rendering api for the loginPageRenderer and Presentation renderer.
The loginPageRenderer is now being called when there is a new presentation. A participant is now removed from the render group when their session expires.  This was the source of a potential memory leak.

Removed PresentationManager code keeping track of sessions to account for active presentations.  There was code calculating and managing expired presentations in the bean - now that the DisposableBean interface is implemented this code is no longer necessary.  

The Presentation's renderer is now disposed when it is empty.  This was the source of another potential memory leak.